Service / Repository 역할 분리

+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFChannelService.java
@@ -16,11 +16,26 @@ public class JCFChannelService implements ChannelService {
 
    @Override
     public void create(Channel channel) {
+       if (channel == null) {
+           throw new IllegalArgumentException("채널 정보가 없습니다.");
+       }
+
+       if (repository.exists(channel.getId())) {
+           throw new IllegalArgumentException("이미 존재하는 채널입니다.");
+       }
         repository.create(channel);
     }
 
     @Override
     public Channel read(UUID id) {
+        if (id == null) {
+            throw new IllegalArgumentException("채널 ID는 필수입니다.");
+        }
+
+        if (!repository.exists(id)) {
+            throw new IllegalArgumentException("존재하지 않는 ID입니다.");
+        }
+
         return repository.read(id);
     }
 
@@ -31,11 +46,28 @@ public class JCFChannelService implements ChannelService {
 
     @Override
     public void update(UUID id, Channel channel) {
+        if (id == null) {
+            throw new IllegalArgumentException("채널 ID는 필수입니다.");
+        }
+
+        if (channel == null) {
+            throw new IllegalArgumentException("수정할 채널 정보가 없습니다.");
+        }
+
+        if (!repository.exists(id)) {
+            throw new IllegalArgumentException("존재하지 않은 ID입니다.");
+        }
         repository.update(id, channel);
     }
 
     @Override
     public void delete(UUID id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID가 존재하지 않습니다.");
+        }
+        if (!repository.exists(id)) {
+            throw new IllegalArgumentException("삭제할 채널이 없습니다.");
+        }
         repository.delete(id);
     }
 }
